+++ src/api/files.js
@@ -77,7 +77,7 @@ export async function isFileExists(fileID) {
  * @param fileID is the id of the file
  */
 export async function getFileByID(fileID) {
-  const res = await Axios.get(`${baseURL}/api/files/${fileID}`);
+  const res = await Axios.get(`${baseURL}/api/files/${fileID}`, {doNotInterfere: true});
   return res.data;
 }
 
@@ -97,7 +97,7 @@ export async function deleteFile(fileID) {
  * multipartUpload create an upload with small size
  * @param file is the file that was chose by the user in the type blob
  */
-export async function multipartUpload({ file, parent }) {
+export async function multipartUpload({ file, parent }, progress) {
   const source = Axios.CancelToken.source();
   const formData = new FormData();
 
@@ -107,13 +107,10 @@ export async function multipartUpload({ file, parent }) {
     formData,
     {
       onUploadProgress: (event) => {
-        store.commit("addLoadingFile", {
-          name: file.name,
-          progress: Math.round((100 * event.loaded) / event.total),
-          source,
-        });
+        progress(file, event, source);
       },
       cancelToken: source.token,
+      doNotInterfere: true,
     }
   );
   const metadata = await getFileByID(res.data);
@@ -124,7 +121,7 @@ export async function multipartUpload({ file, parent }) {
  * resumableUpload is an upload for bigger files
  * @param file is the file to upload
  */
-export async function resumableUpload({ file, parent }) {
+export async function resumableUpload({ file, parent }, progress) {
   const uploadID = await createResumableUpload({ file, parent });
 
   const source = Axios.CancelToken.source();
@@ -136,13 +133,10 @@ export async function resumableUpload({ file, parent }) {
     {
       headers: { "Content-Range": `bytes 0-${file.size - 1}/${file.size}` },
       onUploadProgress: (event) => {
-        store.commit("addLoadingFile", {
-          name: file.name,
-          progress: Math.round((100 * event.loaded) / event.total),
-          source,
-        });
+        progress(file, event, source);
       },
       cancelToken: source.token,
+      doNotInterfere: true,
     }
   );